Thor driver for storr

Storr driver for thor. This allows thor to be used as a storage backend with the storr package and presents a higher level content addressable key/value store suitable for storing R objects.

Usage

storr_thor(
  env,
  prefix = "",
  hash_algorithm = NULL,
  default_namespace = "objects"
)

driver_thor(env, prefix = "", hash_algorithm = NULL)

Arguments

env

A thor environment

prefix

An optional prefix. If given, use a : as the last character for nice looking keys (e.g., storr: will generate keys like storr:keys:namespace:name. If not given then we assume that storr is the only user of this database and if destroy is called it will delete the entire database.

hash_algorithm

Optional hash algorithm to use. Defaults to md5, or whatever the existing algorithm is if the database has been opened. You cannot mix algorithms.

default_namespace

The default namespace to store objects in. Defaults to objects, as does other storr drivers.